requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib
python-multipart
qiskit-ibm-runtime
qiskit
numpy
pandas
scikit-learn
redis
aiofiles
python-dotenv
email-validator
sqladmin
fastapi-sqlalchemy
bcrypt
jupyter
nbformat
nbconvert
httpx
qiskit[all]
bcrypt==4.0.1
passlib==1.7.4
shap
lime
